+++ agency/agency.hpp
@@ -16,7 +16,6 @@
 #include <agency/detail/config.hpp>
 #include <agency/async.hpp>
 #include <agency/bulk_async.hpp>
-#include <agency/bulk_invoke.hpp>
 #include <agency/bulk_then.hpp>
 #include <agency/container.hpp>
 #include <agency/coordinate.hpp>