+++ agency/detail/algorithm/copy/default_copy_n.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/detail/requires.hpp>
-#include <agency/bulk_invoke.hpp>
+#include <agency/functional/bulk_invoke.hpp>
 #include <agency/functional/invoke.hpp>
 #include <agency/execution/execution_policy.hpp>
 #include <agency/detail/type_traits.hpp>