+++ agency/functional.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
+#include <agency/functional/bulk_invoke.hpp>
 #include <agency/functional/invoke.hpp>